+++ src/trio_util/_cancel_scopes.py
@@ -30,3 +30,21 @@ async def move_on_when(fn: Callable[..., Awaitable],
                                cancel_scope.cancel)
             yield cancel_scope
         nursery.cancel_scope.cancel()
+
+
+@asynccontextmanager
+async def run_and_cancelling(fn: Callable[..., Awaitable],
+                             *args, **kwargs) -> AsyncIterator[None]:
+    """Async context manager that runs async fn(*args, **kwargs) and cancels it at block exit.
+
+    Synopsis::
+
+        async with run_and_cancelling(my_background_fn, my_arg=10):
+            await ...
+            # now the block exits, and my_background_fn is cancelled if still running
+    """
+
+    async with trio.open_nursery() as nursery:
+        nursery.start_soon(partial(fn, *args, **kwargs))
+        yield
+        nursery.cancel_scope.cancel()