+++ tests/test_cancel_scopes.py
@@ -2,7 +2,7 @@ import trio
 
 from mock import AsyncMock  # type: ignore[attr-defined]
 
-from trio_util import move_on_when
+from trio_util import move_on_when, run_and_cancelling
 
 
 async def test_move_on_when(autojump_clock):
@@ -39,5 +39,34 @@ async def test_move_on_when_args():
     fn = AsyncMock()
 
     async with move_on_when(fn, 'foo', bar=10):
+        pass
+    fn.assert_awaited_with('foo', bar=10)
+
+
+async def test_run_and_cancelling(autojump_clock):
+    event = trio.Event()
+
+    async def _task():
+        event.set()
+        await trio.sleep_forever()
+
+    async with run_and_cancelling(_task):
+        pass
+        # context manager exits normally and cancels _task
+    assert event.is_set()
+
+    event2 = trio.Event()
+
+    async with run_and_cancelling(trio.sleep, 1):
+        await trio.sleep(2)
+        # background task already exited, this block can still complete
+        event2.set()
+    assert event2.is_set()
+
+
+async def test_run_and_cancelling_args():
+    fn = AsyncMock()
+
+    async with run_and_cancelling(fn, 'foo', bar=10):
         pass
     fn.assert_awaited_with('foo', bar=10)